+++ core/initializer/test_execution/test_executor_parallelizer/test_executor_parallelizer.go
@@ -18,6 +18,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"os"
 	"os/signal"
+	"sort"
 	"sync"
 	"syscall"
 )
@@ -61,9 +62,11 @@ func RunInParallelAndPrintResults(
 	// These need to be buffered else sending to the channel will be blocking
 	testParamsChan := make(chan parallel_test_params.ParallelTestParams, len(allTestParams))
 
+	allTestParamsOrderedKeys := getAllTestParamsOrderedKeys(allTestParams)
+
 	logrus.Debug("Loading test params into work queue...")
-	for _, testParams := range allTestParams {
-		testParamsChan <- testParams
+	for _, allTestParamsKey := range allTestParamsOrderedKeys {
+		testParamsChan <- allTestParams[allTestParamsKey]
 	}
 	close(testParamsChan) // We close the channel so that when all params are consumed, the worker threads won't block on waiting for more params
 	logrus.Debug("All test params loaded into work queue")
@@ -100,7 +103,9 @@ func RunInParallelAndPrintResults(
 	return allTestsPassed
 }
 
-
+// ====================================================================================================
+//                                       Private helper functions
+// ====================================================================================================
 func disableSystemLogAndRunTestThreads(
 		parentContext context.Context,
 		testsuiteExObjNameProvider *object_name_providers.TestsuiteExecutionObjectNameProvider,
@@ -208,4 +213,16 @@ func logErroneousSystemLogging(capturedErroneousMessages []output.ErroneousSyste
 		logrus.StandardLogger().Out.Write(messageInfo.GetStacktrace())
 		logrus.StandardLogger().Out.Write([]byte("\n")) // The stacktrace likely won't end with a newline so we add it
 	}
+}
+/*
+This Helper function receives the allTestParams maps and takes its keys to creates and return an ordered
+slices of strings useful to iterate the allTestParams map in an alphabetical order
+ */
+func getAllTestParamsOrderedKeys(allTestParams map[string]parallel_test_params.ParallelTestParams) []string {
+	allTestParamsOrderedKeys := make([]string, 0, len(allTestParams))
+	for k := range allTestParams {
+		allTestParamsOrderedKeys = append(allTestParamsOrderedKeys, k)
+	}
+	sort.Strings(allTestParamsOrderedKeys)
+	return allTestParamsOrderedKeys
 }